Making some progress, maybe.  This version compiles but doesn't run.
Trying to create a GtkDrawingArea in C++, creating a Cairo context for it,
wrapping the drawing area as a PyGTK widget and passing it out to Python.
Then PyGTK can add it to the window.

testcanvas.py crashes, complaining that GDK_IS_DRAWABLE failed.
Maybe the drawing attempt is coming too soon, before the window
is fully created.