Copied swig infrastructure from OOF2 and hacked the Makefile to use OOF2's swig.
I had been trying to use swig 3.0, but ran into troubles that I don't
feel like solving now.  Since we'll be using OOFCanvas in OOF2, it
makes sense to use OOF2's swig.  I didn't want to copy over OOF2's
swig source code, so the Makefile now explicitly inludes a reference
to the swig in one of *my* OOF2 build directories.  This is suboptimal
but temporary.

testcanvas.py now runs and brings up a Canvas and even draws on it,
although there are problems with the image size.