CanvasItem::draw is not const.  CanvasItem::containsPoint takes a Canvas argument.
CanvasItem::draw can't be const because some items have to update their
bounding boxes when they're drawn, if they don't know their sizes until
then.

CanvasItem::containsPoint has to know the Canvas so that it can convert
between user and device coordinates if necessary.

Added versions of pixel2user and user2pixel that take a single
argument, for converting distances.  This assumes that we aren't
going to support different scales in the x and y directions.