Scaling and scrolling work much better, after major changes.
The size of the Cairo::Surface for each Layer is always adjusted
to fit the bounding boxes of the items in the layer.

Items aren't drawn to the Cairo::Surfaces when they're added to
the layer.  They're now drawn only just before the layer is copied
to the gtk Layout's surface, because the size of the layer isn't
known until then.

Bounding boxes, even for items that need to know the ppu in order
to compute their bounding boxes, are computed before the items' draw
methods are called.

Centering and filling aren't working yet.
Mouse-clicks are broken and disabled.
Text isn't working quite right either.  Maybe the bounding box is wrong.