Refactored so that a C++ version can built without Python.
The Canvas class has been split into CanvasBase and two derived
classes, Canvas and CanvasPython.  Canvas is to be used from C++
and CanvasPython from Python.  CanvasPython is renamed to Canvas
by swig, so the Python interface looks more or less just like the
C++ interface.

The Canvas constructor creates its own gtk layout.  The CanvasPython
constructor expects the layout to be created in Python and passed
in to C++.

The Makefile is ugly but works, I think.  To build the Python version,
use "make PYTHON=1 oofcanvascmodule.so" or just "make PYTHON=1".
The actual value of PYTHON doesn't matter -- it just has to be defined.

To build the C++ version, type "make liboofcanvas.so" on Linux or
"make liboofcanvas.dylib" on Mac.  The libraries seem to build
correctly but I haven't actually used them yet, so this might change.