Added CircleRubberBand, and fixed a few bugs.
Also added a way to make a dashed CanvasSegment, but decided not
to use it with the RubberBands.  Probably all CanvasItems that use
lines of some sort should allow the lines to be dashed, but it must
be possible to set the dash size in a reasonable way.  Cairo wants
to know the lengths in user units, but we don't necessarily know
the right scale.