Creating a new branch to make sure I don't lose track of just what I'm trying to keep....
This branch was created from a detached HEAD when I was backing out the
changes made while trying to separate the gtk and non-gtk code in
the branch 'split-gtk'.  That separation wasn't working because I
couldn't figure out how to build it without the OOF2 build machinery,
which I don't want reproduce here.

This branch was created by checking out the newest commit that didn't
have the split, and used to test a modification that fixed the rubberbands
in OOF2.  I think that I can just merge this into develop, and abandon
split-gtk.