This version actually can be imported into python.
Importing works if done as
>>> import oofcanvas.oofcanvas
>>> import oofcanvas.GUI.oofcanvasgui
which is on the ugly side.  Probably something clever
can be put into the __init__.py files make it easier.
Also, there's an error if oofcanvas.oofcanvas is not
imported before the gui.

It's not possible to stow this version of OOFCanvas and the
gtk3 version of OOF2 at the same time, because OOF2 still
has its built-in oofcanvas module, so this code is basically
untested.  But it can be imported!

Cleaned up setup.py, removing things that were relevant to OOF2
or OOF3D but not OOFCanvas.