oofcanvas.pc is now created by setup.py, via some new distutils hackery.
Distutils is really lacking in some major ways, but switching to
something else at this point seems like a bad idea.  oofcanvas.pc
needs to be created and installed by distutils, using information
from both the build (eg, are we using ImageMagick?) and install
(eg, the prefix) steps.  But if those steps aren't run together,
the data isn't available in one spot.  So I further modified
oof_build_ext so that it writes a version of oofcanvas.pc in the
build temp directory, including placeholders for things it doesn't
know.  Then a modified install_data command copies the file to its
destination after running it through a filter that replaces the
placeholder with the correct value.  The install_data command is
also modified so that its source files can be in the build temp
directory, not just in the source directory.

Added TODOs and comments about some unfinished matters.  New --magick
and --python options to the distutils build command need to be
handled before the set_clib_flags functions in the DIR.py files are
run, but those files are loaded before distutils parses its arguments.