Removed the build args controlling whether the python API is built and ImageMagick is supported.
Now they're controlled by non-distutils options --pythonAPI and
--magick, which are detected and removed from sys.argv before any
distutils calls are made, like many args in oof2 and oof3d.  These
are all hacks, but I couldn't find a better solution.  The problem
is that the DIR.py files have to be read before distutils.core.setup
is called, so that the lists of source files for each Extension can
be constructed.  But if those source files depend on configuration
flags such as --pythonAPI and --magick, the DIR.py need to know the
values of the flags, so they can't be distutils command options.