Fixed an error in oofcanvas.pc.  Temporarily set version number to 0.0.0.
Using oofcanvas.pc when building oof2 means that oofcanvas.pc needs
to contain a real version number, not the placeholder that we've been
using for unreleased development versions.  Until I get the oof2 and
oofcanvas working together, I've set the version to 0.0.0, because
I don't want to deal with packaging releases right now.