The OOFCanvas API doesn't have any explicit references to Cairo or Pango.
This means that OOFCanvas users can just use it without worrying about
any of its under-the-hood details.

Some of the class names and file names are inconsistent and will
be changed.  Files ending with "*impl" or "*private" are implementation
details.  Files ending with "*public" are in the public API.

Canvas items now have two pieces.  The CanvasItem class is visible
to the user and just stores the parameters for the object.  The
details about how to draw it are in a CanvasItemImplementation
class, which is a template, taking the CanvasItem class as the template
parameter.  The implementation is instanced by the CanvasItem constructor.

I undid the previous changes to swiglib, in which I changed the static
names so that they wouldn't conflict with the swiglib objects in OOF2.
That was crashing for some reason.  The simpler solution in any case is
to just wrap all of the swig routines with "namespace OOFCanvas".  The
swig output C++ files already are using that namespace, so nothing
needs to be changed there.

setup.py and DIR.py still need to be changed so that only the
headers for the public API are installed in $prefix/include.

Almost all of this is untested.  But this version compiles, and when
loaded into python CanvasItems can be constructed.