Fixes to bugs discovered when using OOFCanvas within OOF2.
oofcanvas.pc had some errors in it.

Some header files were including cairo and pango headers that
shouldn't be exposed to OOF2.

Changed a RubberBand::start() arg from CanvasLayerImpl*
to CanvasLayer*.

Added a __repr__ for OOFCanvas::Color.

It's currently not possible to add a new RubberBand class in
OOF2, because the type checker in the OOFCanvas swig doesn't
recognize the derived class in OOF2 as a subclass of the base
class in OOFCanvas.  This is a pain and might require a bit of
swig hackery.