Added, and commented out, code to allow classes to be registered after module initialization.
We need to somehow allow a derived class defined in OOF2 to be
passed as a base-class pointer to a function defined in OOFCanvas.
The problem is that when OOFCanvas was swigged, it didn't know
about the derived class, and the typechecker rejects it.

This fix didn't work, for reasons that I don't understand, so I
commented it out.  It may be better in the end to just turn off
swig type checking in OOFCanvas.