Added "-strict 0" to the swig arguments.
This turns off swig type checking, and should allow a RubberBand
subclass defined in OOF2 to be passed as a base class pointer to a
OOFCanvas function that doesn't recognize the subclass.

This works, maybe.  OOF2 crashes sometimes when using the RubberBand,
but I'm not sure if this change is responsible for it.