Added code to maybe make rubberband drawing more efficient.
The code is currently enclosed in #ifdef RESTRICT_RUBBERBAND
blocks, and RESTRICT_RUBBERBAND is not defined.  Comments
say why not.  This may be revisited later, but it might not
be worth the effort.

Also fixed the URL in the comments referring to the gtk docs. The
docs have been reorganized and the old URL was broken.