Starting to add mutex locks in the hope that the intermittent crashes are race conditions.
Canvas and CanvasLayer are mostly locked, but the process is not
complete.

Some functions now have two versions, XXXX and XXXX_nolock.
XXXX acquires a lock and calls XXXX_nolock.  If a function
already has a lock, and needs to call another function, it
can just call the nolock version.