Checkpointing before mucking with refcounts for the wrapped classes.
Swig4 has a different policy for destructors than swig1.1, and we
need to be more careful with wrappers.

The preprocessor OOFCANVAS_USE_PYTHON macro now has a value of 2
or 3 when it's defined.

Commented out the various std::string typemaps.  We should just
use char* variables in the swig files and let swig use its default
behavior.  I think.