Added factory methods for creating CanvasItems, and fixed __repr__s.
When using swig1.1, CanvasItems could be added to a CanvasLayer
from Python via code like
   item = CanvasItemSubclass(args)
   layer.addItem(item)
This worked because the CanvasItem's destructor was not swigged,
so it was not destroyed when the python routine ended.  With
swig4, this seems to be impossible.  If the destructor is not swigged,
then an error message is printed at *run time* about a potential
memory leak, each time the constructor is called.

Now each CanvasItem subclass contains a static create() method,
which returns a pointer to Python but calls the constructor only
in C++, which avoids the problem.  The create() method can be
called from either C++ or Python, so the API still looks the same
(although there's no need to call it from C++).  The constructor
is still available in Python, too, but perhaps it shouldn't be.

I restored some of the std::string typemaps, which apparently are
still needed after all.

A C++ repr(PyObject*) function is now defined in pyutility.h, with
different versions for Python2 and Python3.  It's mostly useful for
debugging.