Fixed some problems with the DEBUG option.
The build type option is now cached, and the DEBUG flag is now
passed to the compiler and to swig if appropriate.  I had thought
that this wasn't necessary since cmake automatically creates the
setting for the build type.

If using swig's -builtin option, it defines SWIGPYTHON_BUILTIN
automatically in swig files.  Now it's defined in C++ too.  We
might need to it when invoking swiglib functions from C++.