Rotated text is now rotated around the right spot.
The definition of "right" depends on whether you think of
origins in the upper left (incorrect) or lower left.

The bounding box for rotated text is still incorrect because I
haven't updated the rotation algorithm for it yet.  But I have
no idea how to correctly compute it even for nonrotated text if
letters with non-zero descents (jyg) are included.