Moved OOFCanvas version info to oofcanvas.h and removed the separate module built from version.swg.
Now the version info is stored in preprocessor macros in C++
(OOFCANVAS_VERSION, OOFCANVAS_VERSION_MAJOR, etc) which are
set by cmake when in builds oofcanvas.h from oofcanvas.h.in.
These macros are swigged as read-only variables, and are available
as oofcanvas.version, oofcanvas.version_major, etc. in python.
This is much cleaner than the old method, and doesn't rely on
an extra script to set the version number.

The ultimate source of the version number is the project() line
at the top of CMakeLists.txt.