A couple changes motivated by inclusion in OOF2.
The generated oofcanvas.pc now includes the ImageMagick flags
in Cflags.

swigruntime.h is protected with "#ifndef SWIGRUNTIME_H", etc.
This isn't sufficient, because the code that is using
OOFCanvas might have its own copy of swigruntime.h which doesn't
have the #ifndef, or has used a different macro name in it.
(Swig is at fault for not including the #ifndef!)  So
pythonexportable.h doesn't includd swigruntime.h
if SWIG_RUNTIME_VERSION is already defined, since
SWIG_RUNTIME_VERSION is defined in swigruntime.h, and it seems
like the sort of thing that we can count on always being there.