Better exception handling.
Borrowed and adapted code from oof2 for transforming C++ exceptions
into python exceptions in the %exception typemap.

When using oofcanvas from python, the exceptiion handling machinery
needs to be initialized by calling init_OOFCanvas(threaded).  The
"threaded" argument is a bool indicating whether or not Python is
using threads (and therefore whether or not the %exception typemap
needs to worry about the global interpreter lock).

Added a single CanvasException class.

Check for exceptions in the draw handler.  These can't be handled
very well, because the handler is called from the main loop.  Now
they print a message and abort the drawing, but don't abort the
main loop.

The one exception that is currently raised is when an attempt is
made to make a new Cairo::ImageSurface that's too big.   This can
happen when zooming in too far.