Slight fixes to GIL handling, and change to the way threading is turned on and off.
This version now works with Python 3.11, although I'm not entirely
sure which change was the important one.  I was having trouble
testing because the oof2 build was sometimes picking up the wrong
version of oofcanvas.pc and therefore using the wrong version of
oofcanvas.   In any case, the routines in pythonlock.h and the
exception typemap in typemaps.swg now are more careful.

threading_enabled is now set with a function call instead of
a simple swigged variable.  The swigged variable wasn't changing
in C++ when set in python.