The ppu for filling the canvas is now found more efficiently.
The code for calculating the pixels-per-unit (ppu) when zooming to
fill the canvas was very slow when large numbers of items were
displayed.  I had thought it was o(N^2), but it was actually o(N^3),
and was prohibitively slow for N > 10000.  The reason for the
slowness is that if there are items defined in device units, instead
of pixel units, the physical size depends on the ppu.   Now if
there are more than LARGE_NUMBER_OF_ITEMS items, a simpler
method is used.  It doesn't quite fill the canvas in some
cases, but it's o(N).