Fixed an error and adding a line to suppress a cmake warning.
OOFCANVAS_USE_PYTHON was being set incorrectly in one spot,
leading to other warnings and possibly errors.

CMake was warning about a policy change in CMake 3.31.  I don't
think it affects us, but I added a cmake_policy() call to enable
the old version of the policy, just to eliminate the warning.